+++ matchbox-frontend/src/app/validate/validation-parameter.ts
@@ -1,11 +1,16 @@
 import {UntypedFormControl} from "@angular/forms";
 
-export class ValidationParameter {
+export class ValidationParameterDefinition {
   param: fhir.r4.OperationDefinitionParameter;
   formControl: UntypedFormControl;
 
   constructor(param: fhir.r4.OperationDefinitionParameter) {
     this.param = param;
     this.formControl = new UntypedFormControl();
+  }
+}
+
+export class ValidationParameter {
+  constructor(readonly name: string, readonly value: string) {
   }
 }